requirements.txt
fastapi>=0.110
uvicorn>=0.27
httpx>=0.27
pyyaml>=6.0
python-multipart>=0.0.9
pydantic>=2.6
